Extract tests into separate files
